requirements.txt
django~=2.2.4
djangorestframework~=3.12.4
django-filter~=2.4.0
